+++ src/mytester.php
@@ -38,6 +38,7 @@ $cmd = new Parser("", [
     "--coverage" => [
         Parser::Argument => true,
         Parser::Optional => true,
+        Parser::Repeatable => true,
     ],
     "--results" => [
         Parser::Argument => true,
@@ -65,7 +66,7 @@ $cmd = new Parser("", [
         Parser::Optional => true,
     ],
 ]);
-/** @var array{path: string, "--colors"?: bool, "--coverage"?: string, "--results"?: string, "--filterOnlyGroups": string, "--filterExceptGroups": string,"--filterExceptFolders": string, "--version"?: bool, "--noPhpt"?: bool} $options */
+/** @var array{path: string, "--colors"?: bool, "--coverage": string[], "--results"?: string, "--filterOnlyGroups": string, "--filterExceptGroups": string,"--filterExceptFolders": string, "--version"?: bool, "--noPhpt"?: bool} $options */
 $options = $cmd->parse();
 
 if (isset($options["--version"])) {
@@ -78,8 +79,8 @@ foreach (CodeCoverageHelper::$defaultEngines as $engine) {
     $codeCoverageCollector->registerEngine(new $engine());
 }
 $codeCoverageCollector->registerFormatter(new PercentFormatter());
-if (isset($options["--coverage"])) {
-    $coverage = explode(":", $options["--coverage"], 2);
+foreach ($options["--coverage"] as $coverage) {
+    $coverage = explode(":", $coverage, 2);
     if (!array_key_exists($coverage[0], CodeCoverageHelper::$availableFormatters)) {
         throw new ValueError("Unknown code coverage formatter " . $coverage[0]);
     }